requirements/dev.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements/dev.txt requirements/base.in requirements/dev.in requirements/quality.in requirements/test.in
#
argparse==1.4.0           # via caniusepython3
astroid==1.6.0            # via pylint
attrs==17.4.0             # via pytest
backports.functools-lru-cache==1.4  # via caniusepython3
caniusepython3==6.0.0
certifi==2017.11.5        # via requests
chardet==3.0.4            # via requests
click==6.7                # via pip-tools
coverage==4.4.2           # via pytest-cov
ddt==1.1.1
diff-cover==1.0.2
distlib==0.2.6            # via caniusepython3
django==2.0.1
factory-boy==2.9.2
faker==0.8.8              # via factory-boy
first==2.0.1              # via pip-tools
futures==3.1.1            # via caniusepython3
idna==2.6                 # via requests
inflect==0.2.5            # via jinja2-pluralize
isort==4.2.15
jinja2-pluralize==0.3.0   # via diff-cover
jinja2==2.10              # via diff-cover, jinja2-pluralize
lazy-object-proxy==1.3.1  # via astroid
markupsafe==1.0           # via jinja2
mccabe==0.6.1             # via pylint
mock==2.0.0
packaging==16.8           # via caniusepython3
path.py==10.5
pbr==3.1.1                # via mock
pep8==1.7.1
pip-tools==1.11.0
pkginfo==1.4.1            # via twine
pluggy==0.6.0             # via pytest, tox
py==1.5.2                 # via pytest, tox
pycodestyle==2.3.1
pydocstyle==2.1.1
pygments==2.2.0           # via diff-cover
pylint==1.8.1
pyparsing==2.2.0          # via packaging
pytest-cov==2.5.1
pytest-django==3.1.2
pytest==3.3.2             # via pytest-cov, pytest-django
python-dateutil==2.6.1    # via faker
pytz==2017.3              # via django
requests-toolbelt==0.8.0  # via twine
requests==2.18.4          # via caniusepython3, requests-toolbelt, twine
six==1.11.0               # via astroid, diff-cover, faker, mock, packaging, pip-tools, pydocstyle, pylint, pytest, python-dateutil, tox
snowballstemmer==1.2.1    # via pydocstyle
text-unidecode==1.1       # via faker
tox-battery==0.5
tox==2.9.1
tqdm==4.19.5              # via twine
twine==1.9.1
urllib3==1.22             # via requests
virtualenv==15.1.0        # via tox
wheel==0.30.0
wrapt==1.10.11            # via astroid
